packages/llamactl/pyproject.toml
[build-system]
requires = ["uv_build>=0.7.20,<0.8.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.3",
  "pytest-cov>=7.0.0",
  "pytest-timeout>=2.4.0",
  "pytest-xdist>=3.8.0",
  "respx>=0.22.0",
  "ty>=0.0.15",
  "ruff>=0.12.9"
]

[project]
name = "llamactl"
version = "0.6.6"
description = "A command-line interface for managing LlamaDeploy projects and deployments"
readme = "README.md"
license = {text = "MIT"}
authors = [
  {name = "Adrian Lyjak", email = "adrianlyjak@gmail.com"}
]
requires-python = ">=3.10, <4"
dependencies = [
  "llama-agents-core[client]>=0.5.0",
  "llama-agents-appserver>=0.5.0",
  "llama-agents-agentcore>=0.5.0",
  "rich>=13.0.0",
  "questionary>=2.0.0",
  "click>=8.2.1",
  "python-dotenv>=1.0.0",
  "tenacity>=9.1.2",
  "textual>=6.0.0",
  "aiohttp>=3.12.14",
  "copier>=9.10.2",
  "pyjwt[crypto]>=2.10.1",
  "typing-extensions>=4.15.0",
  "typing-extensions>=4.15.0 ; python_full_version < '3.11'"
]

[project.optional-dependencies]
agentcore = [
  "boto3>=1.35.0"
]

[project.scripts]
llamactl = "llama_agents.cli:main"

[tool.uv.build-backend]
module-name = ["llama_agents.cli", "llama_deploy.cli"]
namespace = true

[tool.uv.sources]
llama-agents-appserver = {workspace = true}
llama-agents-core = {workspace = true}
llama-agents-agentcore = {workspace = true}
